requirements.txt
# ============================================================================
# DJANGO RESTFUL LOGISTICS TEMPLATE - REQUIREMENTS
# ============================================================================
# Una plantilla profesional de Django REST API para sistemas logísticos
# con arquitectura limpia, inyección de dependencias y mejores prácticas

# ============================================================================
# CORE DJANGO & REST FRAMEWORK
# ============================================================================
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# ============================================================================
# DATABASE & ORM
# ============================================================================
psycopg2-binary==2.9.7
django-model-utils==4.3.1
django-mptt==0.14.0

# ============================================================================
# AUTHENTICATION & SECURITY
# ============================================================================
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7
django-guardian==2.4.0
django-axes==6.1.1
django-ratelimit==4.1.0

# ============================================================================
# API DOCUMENTATION
# ============================================================================
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5
django-silk==5.0.4

# ============================================================================
# ENVIRONMENT & CONFIGURATION
# ============================================================================
python-decouple==3.8
django-environ==0.11.2
pydantic==2.4.2
pydantic-settings==2.0.3

# ============================================================================
# FILE HANDLING & STORAGE
# ============================================================================
Pillow==10.0.1
django-storages==1.14.2
boto3==1.29.7
django-imagekit==4.1.0

# ============================================================================
# CACHING & PERFORMANCE
# ============================================================================
redis==4.6.0
django-redis==5.4.0
django-cachalot==2.6.1
django-cleanup==8.0.0

# ============================================================================
# ASYNC TASKS & BACKGROUND JOBS
# ============================================================================
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.0
flower==2.0.1

# ============================================================================
# WEBSOCKETS & REAL-TIME
# ============================================================================
channels==4.0.0
channels-redis==4.1.0
django-notifications-hq==1.8.3

# ============================================================================
# MONITORING & LOGGING
# ============================================================================
django-extensions==3.2.3
django-debug-toolbar==4.2.0
sentry-sdk==1.38.0
django-health-check==3.17.0

# ============================================================================
# TESTING & QUALITY ASSURANCE
# ============================================================================
pytest==7.4.3
pytest-django==4.5.2
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==19.12.0
coverage==7.3.2
django-nose==1.4.7

# ============================================================================
# CODE QUALITY & FORMATTING
# ============================================================================
black==23.9.1
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0
mypy==1.6.1
django-stubs==4.2.6
djangorestframework-stubs==3.14.4

# ============================================================================
# DEVELOPMENT TOOLS
# ============================================================================
ipython==8.16.1
django-seed==0.3.1
django-silk==5.0.4
django-querycount==0.8.3
memory-profiler==0.61.0

# ============================================================================
# PRODUCTION & DEPLOYMENT
# ============================================================================
gunicorn==21.2.0
whitenoise==6.6.0
uvicorn[standard]==0.24.0
daphne==4.0.0

# ============================================================================
# UTILITIES & HELPERS
# ============================================================================
python-slugify==8.0.1
requests==2.31.0
openpyxl==3.1.2
reportlab==4.0.5
qrcode==7.4.2
python-dateutil==2.8.2
phonenumbers==8.13.24
geopy==2.4.0

# ============================================================================
# DEPENDENCY INJECTION & CLEAN ARCHITECTURE
# ============================================================================
dependency-injector==4.41.0
injector==0.21.0

# ============================================================================
# DATA VALIDATION & SERIALIZATION
# ============================================================================
marshmallow==3.20.1
marshmallow-dataclass==8.6.0
cerberus==1.3.5

# ============================================================================
# ADDITIONAL UTILITIES
# ============================================================================
python-magic==0.4.27
humanize==4.8.0
pytz==2023.3
python-dotenv==1.0.0
